Struct java_properties::PropertiesWriter

pub struct PropertiesWriter<W: Write> { /* fields omitted */ }

Writes to a properties file.

Methods

impl<W: Write> PropertiesWriter<W>

fn new(writer: W) -> Self

Writes to the given Write stream.

fn write_comment(&mut self, comment: &str) -> Result<(), PropertiesError>

Writes a comment to the file.

fn write(&mut self, key: &str, value: &str) -> Result<(), PropertiesError>

Writes a key/value pair to the file.

fn flush(&mut self) -> Result<(), PropertiesError>

Flushes the underlying stream.

fn set_comment_prefix(&mut self, prefix: &str) -> Result<(), PropertiesError>

Sets the comment prefix.

The prefix must contain a '#' or a '!', may only contain spaces, tabs, or form feeds before the comment character, and may not contain any carriage returns or line feeds ('\r' or '\n').

fn set_kv_separator(&mut self, separator: &str) -> Result<(), PropertiesError>

Sets the key/value separator.

The separator may be non-empty whitespace, or a colon with optional whitespace on either side, or an equals sign with optional whitespace on either side. (Whitespace here means ' ', '\t', or '\f'.)

fn set_line_ending(&mut self, line_ending: LineEnding)

Sets the line ending.